+++ packages/tools/components-package/cypress/support/component.d.ts
@@ -1,5 +1,4 @@
 /// <reference types="cypress" />
-import { RenderOptions, HTMLTemplateResult } from 'lit';
 import "cypress-real-events";
 
 export type Renderable = HTMLTemplateResult;
@@ -7,13 +6,13 @@ export interface MountUI5Options extends MountLitTemplateOptions {
 	ui5Configuration: object;
 }
 export type MountOptions = Partial<MountUI5Options>;
-export declare function mount<T extends keyof HTMLElementTagNameMap = any>(component: JSX.Element, options?: MountOptions): Cypress.Chainable<JQuery<HTMLElementTagNameMap[T]>>;
+export declare function mount(component: JSX.Element, options?: MountOptions): CypressChainable<JQuery<HTMLElement>>;
 declare global {
 	namespace Cypress {
 		interface Chainable {
 			/**
 			 * Mount your component into Cypress sandbox
-			 * @param component content to render by lit-html render function
+			 * @param component content to render by preact render function
 			 * @param options render options for custom rendering
 			 */
 			mount: typeof mount;